pyproject.toml
[project]
name = "vervaeke-tutor"
version = "0.1.0"
description = "A Socratic tutor for exploring John Vervaeke's Awakening from the Meaning Crisis"
requires-python = ">=3.8"
dependencies = [
    "anthropic>=0.7.0",
    "python-dotenv>=1.0.0",
    "tiktoken>=0.5.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/vervaeke_tutor"]

[project.scripts]
vervaeke-tutor = "vervaeke_tutor.__main__:main"
